@@ -19,6 +19,7 @@
     "@google-cloud/logging-winston": "^6.0.0",
     "@google-cloud/storage": "^7.0.1",
     "@msgpack/msgpack": "^3.0.0-beta2",
+    "@open-wc/webpack-import-meta-loader": "^0.4.7",
     "better-sqlite3": "^8.4.0",
     "express": "^4.18.2",
     "node-args": "^2.1.8",
@@ -29,8 +30,8 @@
   "packageManager": "yarn@3.6.2",
   "devDependencies": {
     "eslint": "^8.2.0",
-    "webpack": "^5.84",
-    "webpack-cli": "^4.9.1"
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4"
   },
   "browserslist": [
     "last 10 Chrome versions",
